pyproject.toml
[project]
name = "rag-production-starter"
version = "0.1.0"
description = "Production-ready RAG starter kit"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic-settings>=2.1.0",
    "qdrant-client>=1.7.0",
    "openai>=1.10.0",
    "python-dotenv>=1.0.0",
    "ragas>=0.1.0",
    "mlflow>=2.10.0",
    "rank-bm25>=0.2.2",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "markdown-it-py>=3.0.0",
    "tiktoken>=0.5.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.0",
    "ruff>=0.1.0",
    "httpx>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]